moved display functionality into respective components, nixed MovieLibrary
