+++ src/components/VideoStore.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import MovieLibrary from './MovieLibrary';
 import Customer from './Customer';
 import axios from 'axios';
 
@@ -16,32 +15,7 @@ class VideoStore extends Component {
         };
     }
 
-    componentDidMount() {
-        const fullUrl = "http://localhost:3000/customers"
-        axios.get(fullUrl)
-            .then((response) => {
-                const customers = response.data.map((customer) => {
-                    const { id, name, city, address, postal_code, phone, account_credit, movies_checked_out_count } = customer;
-                    const newCustomer = {
-                        id: id,
-                        name: name,
-                        city: city,
-                        address: address,
-                        postal_code: postal_code,
-                        phone: phone,
-                        account_credit: account_credit,
-                        movies_checked_out_count: movies_checked_out_count,
-                    }
-                    return newCustomer;
-                })
 
-                this.setState({ customers });
-
-            })
-            .catch((error) => {
-                this.setState({ errorMessages: error.message });
-            });
-    }
 
     currentCustomerCallback = (id) => {
         this.setState({
@@ -56,22 +30,8 @@ class VideoStore extends Component {
     }
 
     render() {
-        const displayCustomers = this.state.customers.map((customer) => {
-            const { id, name } = customer;
-            return (<Customer id={id} name={name} key={id} />)
-        })
-        return (
 
-            <section className="movielibrary-button">
-                <section className="customer_list">
-                    {displayCustomers}
-                </section>
-                <MovieLibrary
-                    currentMovieCallback={this.currentMovieCallback}
-                    currentCustomerCallback={this.currentCustomerCallback}
-                />
-            </section>
-        );
+        return (<h1>Welcome to our Video Store</h1>)
     }
 }
